WIP #150 decimal price

+++ tools/schema/cart/create_carts.rb
@@ -5,7 +5,7 @@ cql = %Q(CREATE TABLE  carts
             id int Primary Key,
             customer_id bigint,
             currency text,
-            subtotal double,
+            subtotal decimal,
             size int,
             items frozen <map<bigint, map<text,text>>>))
 execute_lazy cql, []
